fix: Only install lazyvim if its not already there

+++ scripts/install.sh
@@ -176,11 +176,13 @@ if [ ! -d "$HOME/.config" ]; then
 fi
 
 # setup lazyvim
-mv ~/.config/nvim{,.bak}
-mv ~/.local/share/nvim{,.bak}
-mv ~/.local/state/nvim{,.bak}
-mv ~/.cache/nvim{,.bak}
-git clone https://github.com/LazyVim/starter ~/.config/nvim
-rm -rf ~/.config/nvim/.git
+if [[ ! -f "$HOME/.config/nvim/lazyvim.json" ]]; then
+  mv ~/.config/nvim{,.bak}
+  mv ~/.local/share/nvim{,.bak}
+  mv ~/.local/state/nvim{,.bak}
+  mv ~/.cache/nvim{,.bak}
+  git clone https://github.com/LazyVim/starter ~/.config/nvim
+  rm -rf ~/.config/nvim/.git
+fi
 
 cd "$OLDPWD" || exit 1
